Fix pickers
